rename sql_query.py to parameterized_query.py
