+++ redash/utils/parameterized_query.py
